+++ crates/nostr-sdk/src/client/mod.rs
@@ -554,13 +554,16 @@ impl Client {
     /// # Ok(())
     /// # }
     /// ```
-    pub async fn subscribe(
+    pub async fn subscribe<F>(
         &self,
-        filter: Filter,
+        filters: F,
         opts: Option<SubscribeAutoCloseOptions>,
-    ) -> Result<Output<SubscriptionId>, Error> {
+    ) -> Result<Output<SubscriptionId>, Error>
+    where
+        F: Into<Vec<Filter>>,
+    {
         let id: SubscriptionId = SubscriptionId::generate();
-        let output: Output<()> = self.subscribe_with_id(id.clone(), filter, opts).await?;
+        let output: Output<()> = self.subscribe_with_id(id.clone(), filters, opts).await?;
         Ok(Output {
             val: id,
             success: output.success,
@@ -578,17 +581,20 @@ impl Client {
     /// It's possible to automatically close a subscription by configuring the [SubscribeAutoCloseOptions].
     ///
     /// Note: auto-closing subscriptions aren't saved in subscriptions map!
-    pub async fn subscribe_with_id(
+    pub async fn subscribe_with_id<F>(
         &self,
         id: SubscriptionId,
-        filter: Filter,
+        filters: F,
         opts: Option<SubscribeAutoCloseOptions>,
-    ) -> Result<Output<()>, Error> {
+    ) -> Result<Output<()>, Error>
+    where
+        F: Into<Vec<Filter>>,
+    {
         let opts: SubscribeOptions = SubscribeOptions::default().close_on(opts);
 
         match &self.gossip {
-            Some(gossip) => self.gossip_subscribe(gossip, id, filter, opts).await,
-            None => Ok(self.pool.subscribe_with_id(id, filter, opts).await?),
+            Some(gossip) => self.gossip_subscribe(gossip, id, filters, opts).await,
+            None => Ok(self.pool.subscribe_with_id(id, filters, opts).await?),
         }
     }
 
@@ -601,19 +607,20 @@ impl Client {
     ///
     /// It's possible to automatically close a subscription by configuring the [SubscribeAutoCloseOptions].
     #[inline]
-    pub async fn subscribe_to<I, U>(
+    pub async fn subscribe_to<I, U, F>(
         &self,
         urls: I,
-        filter: Filter,
+        filters: F,
         opts: Option<SubscribeAutoCloseOptions>,
     ) -> Result<Output<SubscriptionId>, Error>
     where
         I: IntoIterator<Item = U>,
         U: TryIntoUrl,
+        F: Into<Vec<Filter>>,
         pool::Error: From<<U as TryIntoUrl>::Err>,
     {
         let opts: SubscribeOptions = SubscribeOptions::default().close_on(opts);
-        Ok(self.pool.subscribe_to(urls, filter, opts).await?)
+        Ok(self.pool.subscribe_to(urls, filters, opts).await?)
     }
 
     /// Subscribe to filter with custom [SubscriptionId] to specific relays
@@ -622,22 +629,23 @@ impl Client {
     ///
     /// It's possible to automatically close a subscription by configuring the [SubscribeAutoCloseOptions].
     #[inline]
-    pub async fn subscribe_with_id_to<I, U>(
+    pub async fn subscribe_with_id_to<I, U, F>(
         &self,
         urls: I,
         id: SubscriptionId,
-        filter: Filter,
+        filters: F,
         opts: Option<SubscribeAutoCloseOptions>,
     ) -> Result<Output<()>, Error>
     where
         I: IntoIterator<Item = U>,
         U: TryIntoUrl,
+        F: Into<Vec<Filter>>,
         pool::Error: From<<U as TryIntoUrl>::Err>,
     {
         let opts: SubscribeOptions = SubscribeOptions::default().close_on(opts);
         Ok(self
             .pool
-            .subscribe_with_id_to(urls, id, filter, opts)
+            .subscribe_with_id_to(urls, id, filters, opts)
             .await?)
     }
 
@@ -645,15 +653,16 @@ impl Client {
     ///
     /// Subscribe to specific relays with specific filters
     #[inline]
-    pub async fn subscribe_targeted<I, U>(
+    pub async fn subscribe_targeted<I, U, F>(
         &self,
         id: SubscriptionId,
         targets: I,
         opts: SubscribeOptions,
     ) -> Result<Output<()>, Error>
     where
-        I: IntoIterator<Item = (U, Filter)>,
+        I: IntoIterator<Item = (U, F)>,
         U: TryIntoUrl,
+        F: Into<Vec<Filter>>,
         pool::Error: From<<U as TryIntoUrl>::Err>,
     {
         Ok(self.pool.subscribe_targeted(id, targets, opts).await?)
@@ -737,15 +746,18 @@ impl Client {
     ///     .unwrap();
     /// # }
     /// ```
-    pub async fn fetch_events(&self, filter: Filter, timeout: Duration) -> Result<Events, Error> {
+    pub async fn fetch_events<F>(&self, filters: F, timeout: Duration) -> Result<Events, Error>
+    where
+        F: Into<Vec<Filter>>,
+    {
         match &self.gossip {
             Some(gossip) => {
-                self.gossip_fetch_events(gossip, filter, timeout, ReqExitPolicy::ExitOnEOSE)
+                self.gossip_fetch_events(gossip, filters, timeout, ReqExitPolicy::ExitOnEOSE)
                     .await
             }
             None => Ok(self
                 .pool
-                .fetch_events(filter, timeout, ReqExitPolicy::ExitOnEOSE)
+                .fetch_events(filters, timeout, ReqExitPolicy::ExitOnEOSE)
                 .await?),
         }
     }
@@ -758,20 +770,21 @@ impl Client {
     /// To use another exit policy, check [`RelayPool::fetch_events_from`].
     /// For long-lived subscriptions, check [`Client::subscribe_to`].
     #[inline]
-    pub async fn fetch_events_from<I, U>(
+    pub async fn fetch_events_from<I, U, F>(
         &self,
         urls: I,
-        filter: Filter,
+        filters: F,
         timeout: Duration,
     ) -> Result<Events, Error>
     where
         I: IntoIterator<Item = U>,
         U: TryIntoUrl,
+        F: Into<Vec<Filter>>,
         pool::Error: From<<U as TryIntoUrl>::Err>,
     {
         Ok(self
             .pool
-            .fetch_events_from(urls, filter, timeout, ReqExitPolicy::ExitOnEOSE)
+            .fetch_events_from(urls, filters, timeout, ReqExitPolicy::ExitOnEOSE)
             .await?)
     }
 
@@ -846,20 +859,23 @@ impl Client {
     ///
     /// If `gossip` is enabled the events will be streamed also from
     /// NIP65 relays (automatically discovered) of public keys included in filters (if any).
-    pub async fn stream_events(
+    pub async fn stream_events<F>(
         &self,
-        filter: Filter,
+        filters: F,
         timeout: Duration,
-    ) -> Result<BoxedStream<(RelayUrl, Result<Event, Error>)>, Error> {
+    ) -> Result<BoxedStream<(RelayUrl, Result<Event, Error>)>, Error>
+    where
+        F: Into<Vec<Filter>>,
+    {
         match &self.gossip {
             Some(gossip) => {
-                self.gossip_stream_events(gossip, filter, timeout, ReqExitPolicy::ExitOnEOSE)
+                self.gossip_stream_events(gossip, filters, timeout, ReqExitPolicy::ExitOnEOSE)
                     .await
             }
             None => {
                 let stream = self
                     .pool
-                    .stream_events(filter, timeout, ReqExitPolicy::ExitOnEOSE)
+                    .stream_events(filters, timeout, ReqExitPolicy::ExitOnEOSE)
                     .await?;
 
                 // Map stream to change the error type
@@ -879,20 +895,21 @@ impl Client {
     /// To use another exit policy, check [`RelayPool::stream_events_from`].
     /// For long-lived subscriptions, check [`Client::subscribe_to`].
     #[inline]
-    pub async fn stream_events_from<I, U>(
+    pub async fn stream_events_from<I, U, F>(
         &self,
         urls: I,
-        filter: Filter,
+        filters: F,
         timeout: Duration,
     ) -> Result<BoxedStream<(RelayUrl, Result<Event, Error>)>, Error>
     where
         I: IntoIterator<Item = U>,
         U: TryIntoUrl,
+        F: Into<Vec<Filter>>,
         pool::Error: From<<U as TryIntoUrl>::Err>,
     {
         let stream = self
             .pool
-            .stream_events_from(urls, filter, timeout, ReqExitPolicy::default())
+            .stream_events_from(urls, filters, timeout, ReqExitPolicy::default())
             .await?;
 
         // Map stream to change the error type
@@ -909,11 +926,17 @@ impl Client {
     /// This is an **auto-closing subscription** and will be closed automatically on `EOSE`.
     /// To use another exit policy, check [`RelayPool::stream_events_targeted`].
     /// For long-lived subscriptions, check [`Client::subscribe_targeted`].
-    pub async fn stream_events_targeted(
+    pub async fn stream_events_targeted<I, U, F>(
         &self,
-        targets: HashMap<RelayUrl, Filter>,
+        targets: I,
         timeout: Duration,
-    ) -> Result<BoxedStream<(RelayUrl, Result<Event, Error>)>, Error> {
+    ) -> Result<BoxedStream<(RelayUrl, Result<Event, Error>)>, Error>
+    where
+        I: IntoIterator<Item = (U, F)>,
+        U: TryIntoUrl,
+        F: Into<Vec<Filter>>,
+        pool::Error: From<<U as TryIntoUrl>::Err>,
+    {
         let stream = self
             .pool
             .stream_events_targeted(targets, timeout, ReqExitPolicy::default())
@@ -1695,6 +1718,34 @@ impl Client {
         Ok(filters)
     }
 
+    /// Break down filters for gossip and discovery relays
+    async fn break_down_filters<F>(
+        &self,
+        gossip: &GossipWrapper,
+        filters: F,
+    ) -> Result<HashMap<RelayUrl, Vec<Filter>>, Error>
+    where
+        F: Into<Vec<Filter>>,
+    {
+        let filters: Vec<Filter> = filters.into();
+
+        let mut output: HashMap<RelayUrl, HashSet<Filter>> = HashMap::new();
+
+        for filter in filters {
+            let f = self.break_down_filter(gossip, filter).await?;
+
+            for (url, filter) in f {
+                output.entry(url).or_default().insert(filter);
+            }
+        }
+
+        // TODO: avoid this and returns the HashSet. At the moment this is required due to the Into<Vec<Filter>>
+        Ok(output
+            .into_iter()
+            .map(|(k, v)| (k, v.into_iter().collect()))
+            .collect())
+    }
+
     async fn gossip_send_event(
         &self,
         gossip: &GossipWrapper,
@@ -1810,14 +1861,17 @@ impl Client {
         Ok(self.pool.send_event_to(urls, event).await?)
     }
 
-    async fn gossip_stream_events(
+    async fn gossip_stream_events<F>(
         &self,
         gossip: &GossipWrapper,
-        filter: Filter,
+        filters: F,
         timeout: Duration,
         policy: ReqExitPolicy,
-    ) -> Result<BoxedStream<(RelayUrl, Result<Event, Error>)>, Error> {
-        let filters = self.break_down_filter(gossip, filter).await?;
+    ) -> Result<BoxedStream<(RelayUrl, Result<Event, Error>)>, Error>
+    where
+        F: Into<Vec<Filter>>,
+    {
+        let filters = self.break_down_filters(gossip, filters).await?;
 
         // Stream events
         let stream = self
@@ -1832,18 +1886,31 @@ impl Client {
         Ok(Box::pin(stream))
     }
 
-    async fn gossip_fetch_events(
+    async fn gossip_fetch_events<F>(
         &self,
         gossip: &GossipWrapper,
-        filter: Filter,
+        filters: F,
         timeout: Duration,
         policy: ReqExitPolicy,
-    ) -> Result<Events, Error> {
-        let mut events: Events = Events::new(&filter);
+    ) -> Result<Events, Error>
+    where
+        F: Into<Vec<Filter>>,
+    {
+        let filters: Vec<Filter> = filters.into();
+
+        // Construct a new events collection
+        let mut events: Events = if filters.len() == 1 {
+            // SAFETY: this can't panic because the filters are already verified that list isn't empty.
+            let filter: &Filter = &filters[0];
+            Events::new(filter)
+        } else {
+            // More than a filter, so we can't ensure to respect the limit -> construct a default collection.
+            Events::default()
+        };
 
         // Stream events
         let mut stream = self
-            .gossip_stream_events(gossip, filter, timeout, policy)
+            .gossip_stream_events(gossip, filters, timeout, policy)
             .await?;
 
         while let Some((url, result)) = stream.next().await {
@@ -1863,14 +1930,17 @@ impl Client {
         Ok(events)
     }
 
-    async fn gossip_subscribe(
+    async fn gossip_subscribe<F>(
         &self,
         gossip: &GossipWrapper,
         id: SubscriptionId,
-        filter: Filter,
+        filters: F,
         opts: SubscribeOptions,
-    ) -> Result<Output<()>, Error> {
-        let filters = self.break_down_filter(gossip, filter).await?;
+    ) -> Result<Output<()>, Error>
+    where
+        F: Into<Vec<Filter>>,
+    {
+        let filters = self.break_down_filters(gossip, filters).await?;
         Ok(self.pool.subscribe_targeted(id, filters, opts).await?)
     }