+++ crates/nostr-sdk/CHANGELOG.md
@@ -31,6 +31,10 @@
 - Remove `autoconnect` option from `ClientOptions`
 - Change `Client::stream_events*` output to include the `RelayUrl` and `Result`, enabling callers to identify which relay sent the event or if a specific relay encountered an error (https://github.com/rust-nostr/nostr/pull/1156)
 
+### Added
+
+- Re-add support for multi-filter REQ (https://github.com/rust-nostr/nostr/pull/1176)
+
 ## v0.44.1 - 2025/11/09
 
 ### Fixed